style: triangle font and active menu fixes
