+++ apps/carbon/app/modules/parts/ui/Part/PartSidebar/usePartSidebar.tsx
@@ -1,3 +1,4 @@
+import { useParams } from "@remix-run/react";
 import { BiListCheck } from "react-icons/bi";
 import { BsBank, BsBarChartLineFill, BsCartDash } from "react-icons/bs";
 import { HiOutlineCube } from "react-icons/hi";
@@ -6,19 +7,23 @@ import { PiMoneyFill, PiShareNetworkFill } from "react-icons/pi";
 import { usePermissions } from "~/hooks";
 import type { PartReplenishmentSystem } from "~/modules/parts/types";
 import type { Role } from "~/types";
+import { path } from "~/utils/path";
 
 export function usePartSidebar(replenishment: PartReplenishmentSystem) {
   const permissions = usePermissions();
+  const { partId } = useParams();
+  if (!partId) throw new Error("partId not found");
+
   return [
     {
       name: "Details",
-      to: "",
+      to: path.to.partDetails(partId),
       icon: BsBank,
       shortcut: "Command+Shift+d",
     },
     {
       name: "Purchasing",
-      to: "purchasing",
+      to: path.to.partPurchasing(partId),
       isDisabled: replenishment === "Make",
       role: ["employee", "supplier"],
       icon: BsCartDash,
@@ -26,7 +31,7 @@ export function usePartSidebar(replenishment: PartReplenishmentSystem) {
     },
     {
       name: "Suppliers",
-      to: "suppliers",
+      to: path.to.partSuppliers(partId),
       isDisabled: replenishment === "Make",
       role: ["employee", "supplier"],
       icon: PiShareNetworkFill,
@@ -34,7 +39,7 @@ export function usePartSidebar(replenishment: PartReplenishmentSystem) {
     },
     {
       name: "Manufacturing",
-      to: "manufacturing",
+      to: path.to.partManufacturing(partId),
       isDisabled: replenishment === "Buy",
       role: ["employee"],
       icon: BiListCheck,
@@ -42,28 +47,28 @@ export function usePartSidebar(replenishment: PartReplenishmentSystem) {
     },
     {
       name: "Costing",
-      to: "costing",
+      to: path.to.partCosting(partId),
       role: ["employee", "supplier"],
       icon: IoPricetagsOutline,
       shortcut: "Command+Shift+c",
     },
     {
       name: "Planning",
-      to: "planning",
+      to: path.to.partPlanning(partId),
       role: ["employee"],
       icon: BsBarChartLineFill,
       shortcut: "Command+Shift+p",
     },
     {
       name: "Inventory",
-      to: "inventory",
+      to: path.to.partInventory(partId),
       role: ["employee", "supplier"],
       icon: HiOutlineCube,
       shortcut: "Command+Shift+i",
     },
     {
       name: "Sale Price",
-      to: "sale-price",
+      to: path.to.partSalePrice(partId),
       role: ["employee", "customer"],
       icon: PiMoneyFill,
       shortcut: "Command+Shift+x",
