+++ apps/carbon/app/routes/_public+/login.tsx
@@ -85,7 +85,7 @@ export default function LoginRoute() {
         className="hidden dark:block max-w-[100px] mb-3"
       />
 
-      <div className="rounded-lg bg-card shadow-lg p-8 w-[380px]">
+      <div className="rounded-lg bg-card border border-border shadow-lg p-8 w-[380px]">
         <ValidatedForm
           validator={loginValidator}
           defaultValues={{ redirectTo }}
