+++ apps/carbon/app/styles/tailwind.css
@@ -1,4 +1,53 @@
-@import url("https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&family=Roboto+Mono:wght@300;400;500;600;700&display=swap&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@300;400;500;600;700&display=swap&display=swap");
+
+@font-face {
+  font-family: "Triangle";
+  font-style: normal;
+  font-weight: 200;
+  src: url("/fonts/Geist-UltraLight.woff2") format("woff2");
+}
+
+@font-face {
+  font-family: "Triangle";
+  font-style: normal;
+  font-weight: 300;
+  src: url("/fonts/Geist-Light.woff2") format("woff2");
+}
+
+@font-face {
+  font-family: "Triangle";
+  font-style: normal;
+  font-weight: 500;
+  src: url("/fonts/Geist-Medium.woff2") format("woff2");
+}
+
+@font-face {
+  font-family: "Triangle";
+  font-style: normal;
+  font-weight: 600;
+  src: url("/fonts/Geist-SemiBold.woff2") format("woff2");
+}
+
+@font-face {
+  font-family: "Triangle";
+  font-style: normal;
+  font-weight: 700;
+  src: url("/fonts/Geist-Bold.woff2") format("woff2");
+}
+
+@font-face {
+  font-family: "Triangle";
+  font-style: normal;
+  font-weight: 800;
+  src: url("/fonts/Geist-Black.woff2") format("woff2");
+}
+
+@font-face {
+  font-family: "Triangle";
+  font-style: normal;
+  font-weight: 900;
+  src: url("/fonts/Geist-UltraBlack.woff2") format("woff2");
+}
 
 @tailwind base;
 @tailwind components;
